requirements.txt
tensorflow>=2.15
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
scikit-learn>=1.3
seaborn>=0.12
gradio>=4.0
pillow>=10.0
